Refactor board function into cell and utils modules
