+++ src/board/board.ts
@@ -1,3 +1,4 @@
+import { createCell } from "../cell/cell.js";
 import { Board } from "./types";
 
 const createSquareBoard = (dimension: number): Board => {
@@ -5,21 +6,15 @@ const createSquareBoard = (dimension: number): Board => {
     cells: [],
   };
 
-  const generateMine = (): boolean => {
-    const randomNumber = Math.random();
-
-    return randomNumber >= 0.5;
-  };
-
   for (let row = 0; row < dimension; row++) {
     squareBoard.cells.push([]);
+
     for (let column = 0; column < dimension; column++) {
-      squareBoard.cells[row].push({
-        hasMine: generateMine(),
-        minesAround: 0,
-      });
+      squareBoard.cells[row].push(createCell());
     }
   }
 
   return squareBoard;
 };
+
+createSquareBoard(4);
