+++ src/board/types.ts
@@ -1,0 +1,5 @@
+import { Cell } from "../cell/types";
+
+export type Board = {
+  cells: Cell[][];
+};
