+++ src/cell/cell.ts
@@ -1,0 +1,9 @@
+import { generateRandomBoolean } from "../utils/utils.js";
+import { Cell } from "./types";
+
+export const createCell = (): Cell => {
+  return {
+    hasMine: generateRandomBoolean(),
+    minesAround: 0,
+  };
+};
