+++ src/cell/types.ts
@@ -1,0 +1,4 @@
+export type Cell = {
+  hasMine: boolean;
+  minesAround: number;
+};
