--- src/types.ts
@@ -1,8 +1,0 @@
-type Cell = {
-  hasMine: boolean;
-  minesAround: number;
-};
-
-export type Board = {
-  cells: Cell[][];
-};
