+++ src/utils/utils.ts
@@ -1,0 +1,5 @@
+export const generateRandomBoolean = (): boolean => {
+  const randomNumber = Math.random();
+
+  return randomNumber >= 0.7;
+};
